Moved ZeroBuffer Implementation into data-prepper-core and addressed comments
