+++ data-prepper-core/src/main/java/org/opensearch/dataprepper/core/pipeline/SupportsPipelineRunner.java
@@ -1,0 +1,7 @@
+package org.opensearch.dataprepper.core.pipeline;
+
+public interface SupportsPipelineRunner {
+    PipelineRunner getPipelineRunner();
+
+    void setPipelineRunner(PipelineRunner pipelineRunner);
+}
