--- data-prepper-plugins/zero-buffer/build.gradle
@@ -1,27 +1,0 @@
-/*
- * Copyright OpenSearch Contributors
- * SPDX-License-Identifier: Apache-2.0
- */
-
-plugins {
-    id 'java'
-}
-
-dependencies {
-    implementation project(':data-prepper-api')
-    implementation project(':data-prepper-core')
-    implementation 'io.micrometer:micrometer-core'
-    implementation 'org.projectlombok:lombok:1.18.30'
-    annotationProcessor 'org.projectlombok:lombok:1.18.30'
-}
-
-jacocoTestCoverageVerification {
-    dependsOn jacocoTestReport
-    violationRules {
-        rule { //in addition to core projects rule
-            limit {
-                minimum = 0.90
-            }
-        }
-    }
-}
