+++ settings.gradle
@@ -190,5 +190,4 @@ include 'data-prepper-plugins:opensearch-api-source'
 include 'data-prepper-plugins:saas-source-plugins'
 include 'data-prepper-plugins:saas-source-plugins:source-crawler'
 include 'data-prepper-plugins:saas-source-plugins:jira-source'
-include 'data-prepper-plugins:zero-buffer'
